Fixed SVG optimization, switched back from floats to doubles, added stroke cap option

+++ src/main/kotlin/com/maltaisn/mazegen/Configuration.kt
@@ -84,7 +84,7 @@ class Configuration(val mazeSets: List<MazeSet>,
     /**
      * Style settings for drawing the mazes.
      */
-    class Style(val cellSize: Float,
+    class Style(val cellSize: Double,
                 val backgroundColor: Color?,
                 val color: Color,
                 val stroke: BasicStroke,
